Reverted the front-end changes and removed unnecessary functions and files
