+++ templates/task3.html
@@ -5,7 +5,7 @@
     <div class="container">
     <h2>Task 3</h2>
         <hr>
-    <form action="{% url 'ExecuteTask3' %}" method="post" novalidate>
+    <form action="executeTask3" method="post" novalidate>
         {% csrf_token %}
         {{ form|crispy }}
         <button type="submit" class="btn btn-success" name="execute">Execute</button>
